requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
